[UNSTABLE] Type utilities refactor, progress on new expressions
