+++ Libs/Framework/Include/AST/Components/CExprInputs.h
@@ -1,7 +1,9 @@
 // Copyright 2015-2022 Piperift - All rights reserved
 #pragma once
 
+#include "AST/Filtering.h"
 #include "AST/Types.h"
+#include "AST/Utils/Hierarchy.h"
 
 #include <Types/Struct.h>
 
@@ -17,6 +19,15 @@ namespace Rift
 
 		PROP(pinId)
 		AST::Id pinId = AST::NoId;
+
+
+		OutputId() = default;
+		explicit OutputId(TAccessRef<struct CExprOutputs, CChild> access, AST::Id pinId);
+
+		bool IsNone() const
+		{
+			return ::IsNone(nodeId) || ::IsNone(pinId);
+		}
 	};
 
 
@@ -48,5 +59,13 @@ namespace Rift
 			pinIds.Resize(count, AST::NoId);
 			typeIds.Resize(count, AST::NoId);
 		}
+	};
+
+	struct CExprInvalidInputs : public Struct
+	{
+		STRUCT(CExprInvalidInputs, Struct)
+
+		PROP(pinIds)
+		TArray<AST::Id> pinIds;
 	};
 }    // namespace Rift
