+++ Libs/Framework/Include/AST/Utils/TypeUtils.h
@@ -2,14 +2,21 @@
 
 #pragma once
 
+#include "AST/Access.h"
+#include "AST/Components/CChild.h"
+#include "AST/Components/CExprBinaryOperator.h"
+#include "AST/Components/CExprInputs.h"
+#include "AST/Components/CExprOutputs.h"
+#include "AST/Components/CExprUnaryOperator.h"
+#include "AST/Components/CFileRef.h"
+#include "AST/Components/CParent.h"
+#include "AST/Components/CStmtInput.h"
+#include "AST/Components/CStmtOutputs.h"
+#include "AST/Components/CType.h"
+#include "AST/Components/Tags/CChanged.h"
+#include "AST/Components/Tags/CDirty.h"
 #include "AST/Tree.h"
 #include "AST/TypeRef.h"
-
-#include <AST/Access.h>
-#include <AST/Components/CChild.h>
-#include <AST/Components/CFileRef.h>
-#include <AST/Components/CParent.h>
-#include <AST/Components/CType.h>
 
 
 namespace Rift::Types
@@ -18,9 +25,6 @@ namespace Rift::Types
 	Type GetCategory(AST::Tree& ast, AST::Id id);
 
 	AST::Id CreateType(AST::Tree& ast, Type type, Name name = Name::None(), const Path& path = {});
-
-	AST::Id AddVariable(AST::TypeRef type, Name name);
-	AST::Id AddFunction(AST::TypeRef type, Name name);
 
 	void RemoveTypes(TAccessRef<TWrite<CChild>, TWrite<CParent>, CFileRef> access,
 	    TSpan<AST::Id> types, bool removeFromDisk = false);
@@ -34,4 +38,29 @@ namespace Rift::Types
 	bool IsFunctionLibrary(const AST::Tree& ast, AST::Id typeId);
 	bool CanContainVariables(const AST::Tree& ast, AST::Id typeId);
 	bool CanContainFunctions(const AST::Tree& ast, AST::Id typeId);
+
+
+	AST::Id AddVariable(AST::TypeRef type, Name name);
+	AST::Id AddFunction(AST::TypeRef type, Name name);
+
+	AST::Id AddCall(AST::TypeRef type, AST::Id targetFunctionId);
+	AST::Id AddCallInput(AST::Tree& ast, AST::Id functionId, Name name = Name::None());
+	AST::Id AddCallOutput(AST::Tree& ast, AST::Id functionId, Name name = Name::None());
+	void GetCallArgs(AST::Tree& ast, TSpan<AST::Id> callIds, TArray<AST::Id>& inputArgIds,
+	    TArray<AST::Id>& outputArgIds, TArray<AST::Id>& otherIds);
+
+	AST::Id AddIf(AST::TypeRef type);
+	AST::Id AddReturn(AST::TypeRef type);
+
+	AST::Id AddLiteral(AST::TypeRef type, AST::Id literalTypeId);
+	AST::Id AddDeclarationReference(AST::TypeRef type, AST::Id declId);
+	AST::Id AddUnaryOperator(AST::TypeRef type, UnaryOperatorType operatorType);
+	AST::Id AddBinaryOperator(AST::TypeRef type, BinaryOperatorType operatorType);
+
+	AST::Id FindFunctionByName(AST::Tree& ast, Name ownerName, Name functionName);
+
+	using RemoveAccess = TAccess<TWrite<CChanged>, TWrite<CFileDirty>, TWrite<CStmtInput>,
+	    TWrite<CStmtOutputs>, TWrite<CParent>, TWrite<CChild>, CFileRef>;
+	void RemoveNodes(const RemoveAccess& access, TSpan<AST::Id> ids);
+
 }    // namespace Rift::Types
