+++ Libs/Framework/Src/AST/Utils/Expressions.cpp
@@ -2,8 +2,6 @@
 
 #include "AST/Utils/Expressions.h"
 
-#include "AST/Components/CExprInputs.h"
-#include "AST/Components/CExprOutputs.h"
 #include "AST/Types.h"
 #include "AST/Utils/Hierarchy.h"
 
@@ -43,8 +41,7 @@ namespace Rift::AST::Expressions
 
 	bool CanConnect(TAccessRef<CExprInputs, CExprOutputs> access, OutputId output, InputId input)
 	{
-		if (IsNone(output.nodeId) || IsNone(output.pinId) || IsNone(input.nodeId)
-		    || IsNone(input.pinId))
+		if (output.IsNone() || input.IsNone())
 		{
 			return false;
 		}
@@ -71,13 +68,12 @@ namespace Rift::AST::Expressions
 			return false;
 		}
 
-		// Link input
 		auto& inputs = access.Get<CExprInputs>(input.nodeId);
 
+		// Find pin index
 		const i32 index = inputs.pinIds.FindIndex([&input](AST::Id pinId) {
 			return input.pinId == pinId;
 		});
-
 		if (index != NO_INDEX && Ensure(index < inputs.linkedOutputs.Size()))
 		{
 			inputs.linkedOutputs[index] = output;
@@ -86,33 +82,25 @@ namespace Rift::AST::Expressions
 		return false;    // Pin was invalid
 	}
 
-	bool Disconnect(Tree& ast, Id linkId)
+	bool Disconnect(Tree& ast, InputId input)
 	{
-		if (IsNone(linkId))
+		if (input.IsNone() || !ast.Has<CExprInputs>(input.nodeId))
 		{
 			return false;
 		}
 
-		const Id inputPin = linkId;
-		auto* inputComp   = ast.TryGet<CExprInput>(inputPin);
-		if (EnsureMsg(inputComp && !IsNone(inputComp->linkOutputPin),
-		        "Trying to disconnect a unexistant link")) [[likely]]
+		auto& inputs = ast.Get<CExprInputs>(input.nodeId);
+
+		// Find pin index
+		const i32 index = inputs.pinIds.FindIndex([&input](AST::Id pinId) {
+			return input.pinId == pinId;
+		});
+		if (index != NO_INDEX && Ensure(index < inputs.linkedOutputs.Size())) [[likely]]
 		{
-			// We expect the other side to have outputs component
-			auto& outputsComp = ast.Get<CExprOutputs>(inputComp->linkOutputPin);
-			if (Id* lastInputPin = outputsComp.linkInputPins.Find(inputPin)) [[likely]]
-			{
-				*lastInputPin = AST::NoId;
-			}
-			inputComp->linkOutputPin = AST::NoId;
+			OutputId& linked = inputs.linkedOutputs[index];
+			linked           = {};
 			return true;
 		}
 		return false;
-	}
-
-	bool DisconnectFromInputPin(Tree& ast, AST::Id inputPin)
-	{
-		// NOTE: Input pin ids equal input node ids
-		return Disconnect(ast, inputPin);
 	}
 }    // namespace Rift::AST::Expressions
