+++ Tests/AST/Statements.spec.cpp
@@ -2,10 +2,10 @@
 
 #include <AST/FilterAccess.h>
 #include <AST/Tree.h>
-#include <AST/Utils/FunctionUtils.h>
 #include <AST/Utils/Statements.h>
 #include <AST/Utils/TypeUtils.h>
 #include <bandit/bandit.h>
+
 
 
 using namespace snowhouse;
@@ -21,10 +21,10 @@ go_bandit([]() {
 			AST::Id functionId = Types::AddFunction({ast, AST::NoId}, "TestFunction");
 			AssertThat(ast.Has<CStmtOutput>(functionId), Equals(true));
 
-			AST::Id callId = Functions::AddCall({ast, AST::NoId}, functionId);
+			AST::Id callId = Types::AddCall({ast, AST::NoId}, functionId);
 			AssertThat(ast.Has<CStmtOutput>(callId), Equals(true));
 
-			AST::Id ifId = Functions::AddIf({ast, AST::NoId});
+			AST::Id ifId = Types::AddIf({ast, AST::NoId});
 			AssertThat(ast.Has<CStmtOutputs>(ifId), Equals(true));
 			AssertThat(ast.Get<CStmtOutputs>(ifId).linkPins.Size(), Equals(2));
 		});
@@ -35,10 +35,10 @@ go_bandit([]() {
 			AST::Id functionId = Types::AddFunction({ast, AST::NoId}, "TestFunction");
 			AssertThat(ast.Has<CStmtInput>(functionId), Equals(false));
 
-			AST::Id callId = Functions::AddCall({ast, AST::NoId}, functionId);
+			AST::Id callId = Types::AddCall({ast, AST::NoId}, functionId);
 			AssertThat(ast.Has<CStmtInput>(callId), Equals(true));
 
-			AST::Id ifId = Functions::AddIf({ast, AST::NoId});
+			AST::Id ifId = Types::AddIf({ast, AST::NoId});
 			AssertThat(ast.Has<CStmtInput>(ifId), Equals(true));
 		});
 
@@ -46,8 +46,8 @@ go_bandit([]() {
 			AST::Tree ast;
 
 			AST::Id functionId = Types::AddFunction({ast, AST::NoId}, "TestFunction");
-			AST::Id callId     = Functions::AddCall({ast, AST::NoId}, functionId);
-			AST::Id ifId       = Functions::AddIf({ast, AST::NoId});
+			AST::Id callId     = Types::AddCall({ast, AST::NoId}, functionId);
+			AST::Id ifId       = Types::AddIf({ast, AST::NoId});
 
 			AssertThat(AST::Statements::TryConnect(ast, functionId, callId), Equals(true));
 
@@ -67,9 +67,9 @@ go_bandit([]() {
 
 			AST::Id functionId = Types::AddFunction({ast, AST::NoId}, "TestFunction");
 
-			AST::Id ifId    = Functions::AddIf({ast, AST::NoId});
-			AST::Id call1Id = Functions::AddCall({ast, AST::NoId}, functionId);
-			AST::Id call2Id = Functions::AddCall({ast, AST::NoId}, functionId);
+			AST::Id ifId    = Types::AddIf({ast, AST::NoId});
+			AST::Id call1Id = Types::AddCall({ast, AST::NoId}, functionId);
+			AST::Id call2Id = Types::AddCall({ast, AST::NoId}, functionId);
 
 			AssertThat(
 			    AST::Statements::TryConnect(ast, ast.Get<CStmtOutputs>(ifId).linkPins[0], call2Id),
